Add ability to change a project card's "Status" field value (#1)
* Update tests with more specific assertions

The previous assertions were very wide which meant that these test broke
when unrelated changes were made to the function under test.

* Fix mocked response in test

* Extract return type into named type

* Add ability to update "Status" field on added card

This allows the user to specify an override for the default assigned to
all cards that are added to a project via the project's workflows.

This functionality was added to meet an immediate need when migrating
from projects v1 to v2. When writing this, the hope is that in the
future behaviour like this will be supported via project workflows.